requirements.txt
fastapi
pymango
Jinja2
uvicorn
openai
google-generativeai